Decompose snackbars
