+++ packages/talker_flutter/lib/src/ui/widgets/snackbar.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+
+class SnackbarContent extends StatelessWidget {
+  const SnackbarContent({
+    Key? key,
+    required this.message,
+    required this.title,
+    this.bacgroundColor = Colors.red,
+  }) : super(key: key);
+
+  final String title;
+  final String message;
+  final Color bacgroundColor;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(8),
+      decoration: BoxDecoration(
+        color: bacgroundColor,
+        borderRadius: BorderRadius.circular(10),
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          const SizedBox(width: 4),
+          const Icon(Icons.warning, color: Colors.white),
+          const SizedBox(width: 12),
+          Flexible(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text(
+                  title,
+                  style: const TextStyle(
+                    color: Colors.white,
+                    fontSize: 16,
+                    fontWeight: FontWeight.w600,
+                  ),
+                ),
+                Text(
+                  message,
+                  style: const TextStyle(color: Colors.white),
+                ),
+              ],
+            ),
+          ),
+          TextButton(
+            onPressed: () => _closeSnackbar(context),
+            child: const Text("Undo"),
+          )
+        ],
+      ),
+    );
+  }
+
+  void _closeSnackbar(BuildContext context) {
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+  }
+}
